Convert Nexmo::JWT spec to test
